requirements.txt
#python 3.10.0
fastapi
uvicorn
sqlalchemy
passlib[bcrypt]
alembic
bcrypt == 4.0.1
psycopg2-binary
<<<<<<< Updated upstream
httpx
jwt
PyJWT
python-dotenv
python-multipart
=======
postgresql
python-dotenv
httpx
jwt
PyJWT
>>>>>>> Stashed changes
